backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.0.3
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
websockets==12.0
python-multipart==0.0.6
openai==1.3.7
anthropic==0.7.8
azure-cognitiveservices-speech==1.34.0
numpy==1.24.4
scipy==1.11.4
librosa==0.10.1
soundfile==0.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
PyJWT==2.8.0
